requirements.txt
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
aws-cdk-lib==2.129.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
cattrs==23.2.3
click==8.1.7
constructs==10.3.0
fastapi==0.109.2
h11==0.14.0
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
jsii==1.94.0
packaging==23.2
pluggy==1.4.0
publication==0.0.3
py==1.11.0
pydantic==2.6.1
pydantic_core==2.16.2
pytest==6.2.5
python-dateutil==2.8.2
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
toml==0.10.2
typeguard==2.13.3
typing_extensions==4.9.0
uvicorn==0.27.1

# aws-cdk-lib==2.129.0
# constructs>=10.0.0,<11.0.0